ENH: DAO refactored with tests, Service refactored, M:N relationship between Biobank and User, M:N relationship between Project and User

+++ src/main/java/bbmri/DAOimpl/AttachmentDAOImpl.java
@@ -2,15 +2,9 @@ package bbmri.daoImpl;
 
 import bbmri.dao.AttachmentDao;
 import bbmri.entities.Attachment;
-import bbmri.entities.Project;
-import bbmri.entities.User;
 import org.springframework.stereotype.Repository;
 
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
 import java.io.File;
-import java.util.List;
 
 /**
  * Created with IntelliJ IDEA.
@@ -21,45 +15,72 @@ import java.util.List;
  */
 @Repository
 public class AttachmentDaoImpl extends BasicDaoImpl<Attachment> implements AttachmentDao {
-
+ /*
     public void create(Attachment attachment) {
         notNull(attachment);
         em.persist(attachment);
         createFolder(attachment);
     }
-
+    */
+    /*
     public void remove(Attachment attachment) {
         notNull(attachment);
         File file = new File(getPath(attachment));
-        file.delete();
+
+        // Exists?
+        if (!file.exists()) {
+            throw new IllegalArgumentException(
+                    "Delete: no such attachment: " + getPath(attachment));
+        }
+        boolean success = file.delete();
+
+        // Truly deleted?
+        if (!success) {
+            throw new IllegalArgumentException("Delete: deletion failed: " + file);
+        }
+
         File dir = new File(Attachment.ROOT_DIR_PATH + attachment.getProject().getId().toString());
-        if (dir.isDirectory()) {
-            if (dir.list().length < 1) {
-                dir.delete();
+
+        // Correct path to parent directory?
+        if (!dir.exists()) {
+            throw new IllegalArgumentException(
+                    "Delete: no such folder: " + dir);
+        }
+        // It is truly a directory?
+        if (!dir.isDirectory()) {
+            throw new IllegalArgumentException(
+                    "Delete: not a directory: " + dir);
+        }
+
+        // If empty - delete it!
+        if (dir.list().length < 1) {
+            success = dir.delete();
+            if (!success) {
+                throw new IllegalArgumentException("Delete: deletion failed: " + dir);
             }
         }
         em.remove(attachment);
     }
+    */
 
+   /*
     private void createFolder(Attachment attachment) {
         notNull(attachment);
         File rootDir = new File(Attachment.ROOT_DIR);
         if (!rootDir.exists()) {
-            rootDir.mkdir();
+            boolean success = rootDir.mkdir();
+            if (!success) {
+                throw new IllegalArgumentException("Create folder: failed: " + rootDir);
+            }
         }
         File dir = new File(Attachment.ROOT_DIR_PATH
                 + attachment.getProject().getId().toString());
         if (!dir.exists()) {
-            dir.mkdir();
+            boolean success =  dir.mkdir();
+            if (!success) {
+                throw new IllegalArgumentException("Create folder: failed: " + dir);
+            }
         }
     }
-
-    public String getPath(Attachment attachment) {
-        notNull(attachment);
-        return (Attachment.ROOT_DIR_PATH
-                + attachment.getProject().getId().toString()
-                + File.separator
-                + attachment.getProject().getId().toString()
-                + attachment.getAttachmentType().toString());
-    }
+    */
 }
